pyproject.toml
[project]
name = "sbs-python"
version = "0.1.0"
description = "Somebody-Skills Python 蒸馏引擎"
requires-python = ">=3.10"

dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.34.0",
  "sentence-transformers>=3.4.0",
  "chromadb>=0.6.3",
  "hdbscan>=0.8.40",
  "torch>=2.6.0",
  "transformers>=4.48.0",
  "llama-cpp-python>=0.3.0",
  "jinja2>=3.1.5",
  "tiktoken>=0.8.0",
  "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"
